refactor(dataset): async moved to its own module
Plus, some shared constants across module is moved in a dedicated module.
