+++ src/hp_pred/constants.py
@@ -1,0 +1,1 @@
+VITAL_API_BASE_URL = "https://api.vitaldb.net"
